requirements.txt
absl-py==2.1.0
accelerate==0.27.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.6.1
apiclient==1.0.4
appdirs==1.4.4
asgiref==3.8.1
astunparse==1.6.3
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
beautifulsoup4==4.12.2
black==22.10.0
blinker==1.7.0
boto3==1.34.39
botocore==1.34.39
bpemb==0.3.3
bs4==0.0.1
build==1.0.3
CacheControl==0.13.1
cachetools==4.2.4
certifi==2023.11.17
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
clang==5.0
cleo==2.1.0
click==8.1.7
cloudpickle==1.6.0
colorama==0.4.4
conllu==4.5.3
crashtest==0.4.1
cycler==0.10.0
databases==0.9.0
dataclasses-json==0.6.4
decorator==4.4.2
Deprecated==1.2.14
distlib==0.3.7
distro==1.9.0
dulwich==0.21.6
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fastapi==0.109.2
fastjsonschema==2.19.0
filelock==3.13.1
flair==0.13.1
Flask==2.2.5
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.1.3
future==0.18.2
gast==0.4.0
gdown==5.1.0
gensim==4.3.2
google-api-core==2.11.0
google-api-python-client==2.72.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
googleapis-common-protos==1.58.0
greenlet==3.0.1
grpcio==1.60.1
h11==0.13.0
h5py==3.1.0
httpcore==1.0.2
httplib2==0.21.0
httpx==0.25.2
huggingface-hub==0.20.3
hyperopt==0.2.5
idna==2.10
importlib-metadata==7.0.0
installer==0.7.0
itsdangerous==2.1.2
Janome==0.5.0
jaraco.classes==3.3.0
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
keras==2.15.0
Keras-Preprocessing==1.1.2
keyring==24.3.0
kiwisolver==1.3.1
konoha==4.6.5
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.46
langchain-postgres==0.0.3
langchain-text-splitters==0.0.1
langdetect==1.0.9
langgraph==0.0.32
langsmith==0.1.52
libclang==16.0.6
loguru==0.7.2
lxml==5.1.0
Markdown==3.3.4
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.4.2
ml-dtypes==0.2.0
more-itertools==10.1.0
mpld3==0.3
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.5
mypy-extensions==1.0.0
networkx==2.5.1
nltk==3.6.3
numpy==1.22.4
oauth2client==4.1.3
oauthlib==3.1.1
openai==1.17.1
opencv-python==4.6.0.66
openpyxl==3.0.7
opt-einsum==3.3.0
orjson==3.10.1
outcome==1.3.0.post0
overrides==3.1.0
packaging==23.2
pandas==1.3.0
pathspec==0.10.2
pexpect==4.9.0
pgvector==0.2.5
Pillow==9.2.0
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
pptree==3.1
protobuf==4.21.12
psutil==5.9.8
psycopg==3.1.18
psycopg-pool==3.2.1
psycopg2==2.9.9
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.11
pydantic_core==2.16.2
pyparsing==2.4.7
pyproject_hooks==1.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.5
pytorch_revgrad==0.2.0
pytz==2021.1
pywin32-ctypes==0.2.2
PyYAML==5.4.1
qrcode==7.3.1
rapidfuzz==3.5.2
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.0
requests-toolbelt==1.0.0
rsa==4.7.2
s3transfer==0.10.0
sacremoses==0.0.45
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.7.0
segno==1.5.2
segtok==1.5.11
selenium==4.15.2
semver==3.0.2
sentence-transformers==2.3.1
sentencepiece==0.1.99
shellingham==1.5.4
simplejson==3.17.3
six==1.15.0
smart-open==5.1.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.29
sqlitedict==2.1.0
starlette==0.36.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.0
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0      
termcolor==1.1.0
threadpoolctl==2.2.0
tiktoken==0.6.0
tokenizers==0.15.1
tomli==2.0.1
tomlkit==0.12.3
torch==2.2.0
tqdm==4.66.2
transformer-smaller-training-vocab==0.3.3 
transformers==4.37.2
trio==0.23.1
trio-websocket==0.11.1
trove-classifiers==2023.11.29
typing-inspect==0.8.0
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.6
uvicorn==0.18.2
virtualenv==20.4.7
watchfiles==0.16.1
wcwidth==0.2.13
websockets==10.3
Werkzeug==3.0.2
Wikipedia-API==0.6.0
win32-setctime==1.1.0
wrapt==1.12.1
wsproto==1.2.0
yarl==1.9.4
zipp==3.5.0